download ssl checkpoint if necessary
